IMP(Create Namespace Command): Build a CloudFormation stack instead of creating loose resources in AWS
