+++ package.json
@@ -8,7 +8,7 @@
   "private": false,
   "scripts": {
     "clean": "rm -rf ./dist",
-    "build:copy": "cpx 'src/**/*.yaml' dist",
+    "build:copy": "cpx 'src/**/*.{yaml,json}' dist",
     "build:ts": "tsc",
     "build": "yarn clean && yarn build:copy && yarn build:ts",
     "start": "node dist/index.js",
@@ -19,28 +19,24 @@
   "dependencies": {
     "@kubernetes/client-node": "^0.7.2",
     "@lpha/core": "^0.3.16",
-    "aws-sdk": "^2.382.0",
-    "js-yaml": "^3.12.0",
+    "aws-sdk": "^2.441.0",
+    "js-yaml": "^3.13.1",
     "lodash": "^4.17.11"
   },
   "devDependencies": {
-    "@types/js-yaml": "^3.11.4",
-    "@types/lodash": "^4.14.119",
-    "@types/node": "^10.12.18",
+    "@types/js-yaml": "^3.12.1",
+    "@types/lodash": "^4.14.123",
+    "@types/node": "^11.13.7",
     "@types/request": "^2.48.1",
     "@types/ws": "^6.0.1",
     "cpx": "^1.5.0",
-    "tslint": "^5.12.0",
-    "tslint-config-airbnb-base": "^0.2.0",
-    "typescript": "^3.2.2"
+    "tslint": "^5.16.0",
+    "tslint-config-airbnb-base": "^0.3.0",
+    "typescript": "^3.4.4"
   },
   "@": {
     "commands": [
       "./dist"
-    ],
-    "setup-cluster": {
-      "name": "boldare",
-      "region": "eu-west-1"
-    }
+    ]
   }
 }
