+++ src/commands/create-namespace/index.ts
@@ -1,0 +1,1 @@
+export * from './create-namespace.command';
