+++ src/commands/index.ts
@@ -2,7 +2,6 @@ export * from './install-autoscaler';
 export * from './install-dashboard';
 export * from './install-dns';
 export * from './create-admin-user.command';
-export * from './create-namespace.command';
-export * from './create-namespace-user.command';
+export * from './create-namespace';
 export * from './launch-dashboard.command';
 export * from './setup-cluster.command';
